app/requirements.txt
numpy
uvicorn
Pillow
onnxruntime
fastapi
pydantic
python-multipart
google-cloud-storage
requests
sqlalchemy